Added changelog display for updates
Also incremented cache version to ignore old cache, and cleaned up config menu a bit.
