+++ package.json
@@ -15,6 +15,7 @@
     "spicetify-creator": "^1.0.17"
   },
   "dependencies": {
-    "idb": "^8.0.3"
+    "idb": "^8.0.3",
+    "marked": "^17.0.1"
   }
 }
