+++ src/types/versionInfo.ts
@@ -1,0 +1,8 @@
+type VersionInfo = {
+  isOutdated: boolean;
+  latestVersion: string;
+  downloadUrl: string;
+  changelog: string;
+};
+
+export { VersionInfo }
